pyproject.toml
[project]
name = "fathom-financial-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "grandalf>=0.8",
    "jupyter>=1.1.1",
    "langgraph>=1.0.6",
    "llama-index>=0.14.12",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-together>=0.4.1",
    "llama-index-vector-stores-postgres>=0.7.2",
    "llama-parse>=0.6.54",
    "modal>=1.3.1",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "peft>=0.18.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "together>=1.5.35",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.5",
]
